rattler_virtual_packages::win

Function windows_version

pub fn windows_version() -> Option<Version>

Returns the Windows version of the current platform.

Returns an error if determining the Windows version resulted in an error. Returns None if the Windows version could not be determined. Note that this does not mean the current platform is not Windows.